+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/ServiceOwner/DialogTransmissions/Create/CreateDialogTransmissionEndpoint.cs
@@ -35,6 +35,7 @@ public sealed class CreateDialogTransmissionEndpoint : Endpoint<CreateTransmissi
             StatusCodes.Status201Created,
             StatusCodes.Status400BadRequest,
             StatusCodes.Status404NotFound,
+            StatusCodes.Status410Gone,
             StatusCodes.Status412PreconditionFailed,
             StatusCodes.Status422UnprocessableEntity));
     }
@@ -67,7 +68,7 @@ public sealed class CreateDialogTransmissionEndpoint : Endpoint<CreateTransmissi
         await result.Match(
             success => SendCreatedAtAsync<GetDialogTransmissionEndpoint>(new GetTransmissionQuery { DialogId = dialog.Id, TransmissionId = req.Id.Value }, req.Id, cancellation: ct),
             notFound => this.NotFoundAsync(notFound, ct),
-            badRequest => this.BadRequestAsync(badRequest, ct),
+            gone => this.GoneAsync(gone, ct),
             validationError => this.BadRequestAsync(validationError, ct),
             forbidden => this.ForbiddenAsync(forbidden, ct),
             domainError => this.UnprocessableEntityAsync(domainError, ct),