+++ src/Digdir.Domain.Dialogporten.WebApi/Endpoints/V1/ServiceOwner/DialogTransmissions/Get/GetDialogTransmissionEndpoint.cs
@@ -24,7 +24,8 @@ public sealed class GetDialogTransmissionEndpoint : Endpoint<GetTransmissionQuer
 
         Description(b => b.ProducesOneOf<TransmissionDto>(
             StatusCodes.Status200OK,
-            StatusCodes.Status404NotFound));
+            StatusCodes.Status404NotFound,
+            StatusCodes.Status410Gone));
     }
 
     public override async Task HandleAsync(GetTransmissionQuery req, CancellationToken ct)
@@ -32,6 +33,7 @@ public sealed class GetDialogTransmissionEndpoint : Endpoint<GetTransmissionQuer
         var result = await _sender.Send(req, ct);
         await result.Match(
             dto => SendOkAsync(dto, ct),
-            notFound => this.NotFoundAsync(notFound, ct));
+            notFound => this.NotFoundAsync(notFound, ct),
+            deleted => this.GoneAsync(deleted, ct));
     }
 }